# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Haife API",
    "version": "1.0.0",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/user": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "新增使用者",
        "description": "傳入使用者資訊並建立使用者",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "description": "使用者權杖",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "使用者新增資訊",
            "schema": {
              "type": "object",
              "properties": {
                "role": {
                  "type": "string",
                  "example": "使用者角色"
                },
                "realName": {
                  "type": "string",
                  "example": "真實姓名"
                },
                "nickName": {
                  "type": "string",
                  "example": "暱稱"
                },
                "account": {
                  "type": "string",
                  "example": "帳號"
                },
                "password": {
                  "type": "string",
                  "example": "密碼"
                },
                "isEnabled": {
                  "type": "string",
                  "example": "是否啟用"
                }
              },
              "required": [
                "role",
                "realName",
                "nickName",
                "account",
                "password",
                "isEnabled"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "新增成功"
          },
          "400": {
            "description": "新增失敗"
          },
          "403": {
            "description": "使用者沒有角色頁面權限"
          }
        }
      },
      "get": {
        "tags": [
          "User"
        ],
        "summary": "使用者查詢 (含分頁與搜尋)",
        "description": "列出所有使用者資訊 (含分頁與搜尋)",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "description": "使用者權杖",
            "type": "string"
          },
          {
            "name": "rowsPerPage",
            "in": "query",
            "required": false,
            "description": "每頁筆數",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "目前頁數",
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "description": "排序欄位",
            "type": "string"
          },
          {
            "name": "sortType",
            "in": "query",
            "required": false,
            "description": "排序方式",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "description": "搜尋資訊 (可空值)",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "查詢成功"
          },
          "400": {
            "description": "查詢失敗"
          },
          "403": {
            "description": "使用者沒有角色頁面權限"
          }
        }
      }
    },
    "/api/user/{uuid}": {
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "修改使用者",
        "description": "根據使用者編號，傳入使用者資訊並修改使用者",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "使用者編號"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "description": "使用者權杖",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "使用者修改資訊",
            "schema": {
              "type": "object",
              "properties": {
                "role": {
                  "type": "string",
                  "example": "使用者角色"
                },
                "realName": {
                  "type": "string",
                  "example": "真實姓名"
                },
                "nickName": {
                  "type": "string",
                  "example": "暱稱"
                },
                "account": {
                  "type": "string",
                  "example": "帳號"
                },
                "password": {
                  "type": "string",
                  "example": "密碼"
                },
                "isEnabled": {
                  "type": "string",
                  "example": "是否啟用"
                }
              },
              "required": [
                "role",
                "realName",
                "nickName",
                "account",
                "password",
                "isEnabled"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "修改成功"
          },
          "400": {
            "description": "修改失敗"
          },
          "403": {
            "description": "使用者沒有角色頁面權限"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "刪除使用者",
        "description": "根據使用者編號刪除使用者",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "使用者編號"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "description": "使用者權杖",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "刪除成功"
          },
          "400": {
            "description": "刪除失敗"
          },
          "403": {
            "description": "使用者沒有角色頁面權限"
          }
        }
      }
    },
...
    "/api/winners/{uuid}": {
      "patch": {
        "tags": [
          "Winners"
        ],
        "summary": "修改中獎人",
        "description": "根據中獎人編號，傳入中獎資訊並修改中獎人",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "中獎人編號"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "description": "使用者權杖",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "中獎人修改資訊",
            "schema": {
              "type": "object",
              "properties": {
                "isExchange": {
                  "type": "string",
                  "example": "是否兌換"
                },
                "remark": {
                  "type": "string",
                  "example": "備註"
                }
              },
              "required": [
                "isExchange",
                "remark"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "修改成功"
          },
          "400": {
            "description": "修改失敗"
          },
          "403": {
            "description": "使用者沒有角色頁面權限"
          }
        }
      }
    },
    "/api/winners/{activity_uuid}": {
      "get": {
        "tags": [
          "Winners"
        ],
        "summary": "中獎人查詢 (含分頁與搜尋)",
        "description": "根據活動編號列出所有中獎人資訊 (含分頁與搜尋)",
        "parameters": [
          {
            "name": "activity_uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "活動編號"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "description": "使用者權杖",
            "type": "string"
          },
          {
            "name": "rowsPerPage",
            "in": "query",
            "required": false,
            "description": "每頁筆數",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "目前頁數",
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "description": "排序欄位",
            "type": "string"
          },
          {
            "name": "sortType",
            "in": "query",
            "required": false,
            "description": "排序方式",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "description": "搜尋資訊 (可空值)",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "查詢成功"
          },
          "400": {
            "description": "查詢失敗"
          },
          "403": {
            "description": "使用者沒有角色頁面權限"
          }
        }
      }
    },
...
  }
}